# src/api/swagger.json
{
    "openapi": "3.0.0",
    "servers": [],
    "info": {
        "title": "Big Data & Moi — API",
        "description": "API de l'application BigDataEtMoi réalisé en NodeJS, Express, TypeScript et MongoDB.",
        "version": "1.0.0"
    },
    "tags": [{
            "name": "Info"
        },
        {
            "name": "User"
        },
        {
            "name": "Wifi"
        },
        {
            "name": "Message"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Info"
                ],
                "summary": "Show API information",
                "description": "This is a public route",
                "operationId": "showApiInfo",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List all users",
                "description": "Returns all users",
                "operationId": "FindAllUsers",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "description": "",
                "operationId": "CreateUser",
                "responses": {
                    "200": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "bad request, body not valid"
                    },
                    "409": {
                        "description": "email address already registered"
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/NewUser"
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "description": "Returns the given user",
                "operationId": "FindUser",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to return",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "description": "Updates the given user",
                "operationId": "UpdateUser",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to update",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "bad request, body not valid"
                    },
                    "404": {
                        "description": "user not found"
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/NewUser"
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "description": "Removes the given user",
                "operationId": "DeleteUser",
                "security": [{
                    "basicAuth": []
                }],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to delete",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/wifi": {
            "post": {
                "tags": [
                    "Wifi"
                ],
                "summary": "Add wifi to user which post the request",
                "description": "Create wifi",
                "operationId": "CreateWifi",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserWithWifi"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "bad request"
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/NewWifi"
                }
            }
        },
        "/messages": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Add message to user which post the request",
                "description": "Create message",
                "operationId": "CreateMessage",
                "security": [{
                    "basicAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserWithMessages"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "bad request"
                    }
                },
                "requestBody": {
                    "$ref": "#/components/requestBodies/NewMessage"
                }
            }
        }
    },
    "components": {
        "requestBodies": {
            "NewUser": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NewUser"
                        }
                    }
                },
                "description": "User object that needs to be added to the database",
                "required": true
            },
            "NewWifi": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/Wifi"
                            }
                        }
                    }
                },
                "description": "Wifi array that needs to be added to the user",
                "required": true
            },
            "NewMessage": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        },
        "schemas": {
            "NewUser": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "title": "User",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uri"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "wifiList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wifi"
                        }
                    },
                    "messageList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        }
                    }
                }
            },
            "Wifi": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "ssid": {
                        "type": "string"
                    }
                }
            },
            "Message": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    }
                }
            },
            "UserWithWifi": {
                "title": "UserWithWifi",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uri"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "wifiList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wifi"
                        }
                    }
                }
            },
            "UserWithMessages": {
                "title": "UserWithMessages",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uri"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "messageList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        }
                    }
                }
            }
        }
    }
}